Fire Pile
Added Fire Pile with 3 Stages as an early Game Substitute for Netherrack.
Fire Piles burn for a limited Amount of Time and need to be refilled with Planks.
Fire Piles can be lit using Torches or Flint and Steel.
Depleted Fire Piles can be reloaded with Planks.
Breaking a Fire Pile drops Coal Dust.